apply partial grid also at grounded margins.
This extends the partial grid approach also to grounded margins.
It applies when part_grid is on.
The floating condition for an ice thickness that would extend
the adjacent ice boxes with zero surface gradient
determines if the partial grid cell is floating or grounded.
Cliffs ((average surface height minus average ice thickness) > 0)
are excluded.