subgrid interpolation of basal melting in iMenergy and iMenthalpy.
the subgrid interpolation for basal melting uses gl_mask to apply basal melting
as a ratio of grounded and floating basal melt to grid boxes at the grounding line.
this influences bwat in iMenthalpy.cc, so strong ocean melt leads to reduced
bed friction at the grounding line of marine based glaciers.
This is not the case for cold mode (see iMtemp.cc).

This is a combination of 3 commits.
The first commit's message is:
remove unneeded glmask access in iMgeometry.cc

This is the 2nd commit message:
subgrid melting when enthalpy model is used.

This is the 3rd commit message:
make sure that sub-grid interp. of basal melt also is applied in cold mode